Structure your data
In this lab, you will practice structuring, an integral step in exploratory data analysis that helps you find and use the right tools to adjust and filter datasets and problem-solve. You will be presented with a business scenario and a dataset to use as you explore structuring. You will practice applying data structuring and using visualizations to analyze data and develop your knowledge of functions that allow you to structure data.